jd_api/requirements.txt
anyio==3.6.2
azure.storage.blob==12.17.0
beautifulsoup4==4.8.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
confection==0.0.3
configobj==5.0.6
configparser==5.3.0
cymem==2.0.7
etelemetry==0.3.0
fastapi==0.86.0
filelock==3.8.0
future==0.18.2
h11==0.14.0
httplib2==0.21.0
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
langcodes==3.3.0
looseversion==1.0.2
lxml==4.9.1
MarkupSafe==2.1.1
murmurhash==1.0.9
networkx==2.8.8
nibabel==4.0.2
nipype==1.8.5
numpy==1.23.4
packaging==21.3
pandas==1.4.4
pathlib==1.0.1
pathy==0.6.2
Pillow==9.3.0
preshed==3.0.8
prov==2.0.0
pydantic==1.9.2
pydot==1.4.2
PyMuPDF==1.19.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytz==2022.6
pyxnat==1.5
rdflib==6.2.0
requests==2.28.1
scipy==1.9.3
sentry-sdk==1.10.1
simplejson==3.17.6
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-transformers==1.1.8
srsly==2.4.5
starlette==0.20.4
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.13.0
thinc==8.1.5
tqdm==4.64.1
traits==6.3.2
typer==0.4.2
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
wasabi==0.10.1
